Create Encoder 4 to 2 Design File
A 4 to 2 encoder is a digital device that takes four input lines and produces a two-bit binary code on two output lines.

Operation: Only one of the four inputs can be "1" at a time. The output code is determined by the combination of input lines. 

Applications: The 4 to 2 encoder is used in systems where fewer output lines are needed, such as for simple data compression.